Fix LIBS_BATS_MOCK_VERSION version number
There's a convenient `LIBS_BATS_MOCK_VERSION` environment variable that's set in the `Dockerfile` that's also happens to be a complete lie. The version set in the environment is `1.1.0` but the version actually installed in `1.3.0`.

This commit makes the version consistent by leveraging the environment variable at installation time, so this never is a problem again.